mix msvc build
msvc doesn't accept NULL/0 in the std::function constructor

+++ src/intro.cpp
@@ -53,7 +53,7 @@ void the_end(CVideo &video, std::string text, unsigned int duration)
 		if(n)
 			sdl::fill_rect(video.getSurface(),&area,0);
 
-		const SDL_Color col = create_color(n, n, n, n);
+		const SDL_Color col = create_color(uint8_t(n), uint8_t(n), uint8_t(n), uint8_t(n));
 		font::draw_text(&video,area,font_size,col,text,area.x,area.y);
 		update_rect(area);
